Improve icon colors for the IntelliJ tool window icons
Without this fix, the continue icon sticks out from the other toolwindow icons, resulting in an inconsistent appearing and creates a feeling that "it doesn't fit into the IDE".

According to
https://plugins.jetbrains.com/docs/intellij/icons.html#new-ui-icon-colors
specific colors are needed to work nicely with dark and light modes. Bonus is that the active tool window icon color then changes automatically to white.